Moving CsvParser to format/
