+++ src/CMakeLists.txt
@@ -37,7 +37,6 @@ set(keepassx_SOURCES
         core/Bootstrap.cpp
         core/Clock.cpp
         core/Config.cpp
-        core/CsvParser.cpp
         core/CustomData.cpp
         core/Database.cpp
         core/DatabaseIcons.cpp
@@ -71,6 +70,7 @@ set(keepassx_SOURCES
         crypto/kdf/AesKdf.cpp
         crypto/kdf/Argon2Kdf.cpp
         format/CsvExporter.cpp
+        format/CsvParser.cpp
         format/HtmlExporter.cpp
         format/KeePass1Reader.cpp
         format/KeePass2.cpp
