+++ src/gui/csvImport/CsvParserModel.h
@@ -21,8 +21,8 @@
 
 #include <QAbstractTableModel>
 
-#include "core/CsvParser.h"
 #include "core/Group.h"
+#include "format/CsvParser.h"
 
 class CsvParserModel : public QAbstractTableModel, public CsvParser
 {
